navbar button perf based debounce
