+++ src/components/canvas/navbar/single-button.tsx
@@ -2,6 +2,12 @@ import { useState, useEffect } from "react";
 import * as LucideIcons from "lucide-react";
 import { AnimatePresence, motion } from "framer-motion";
 import { useToast } from "../../hooks/use-toast";
+import { usePerformanceMode } from "~/hooks/usePerformanceMode";
+import { copyText } from "~/lib/copy";
+
+// Shared global cooldown state so all SingleButton instances share debounce
+let globalBlocked = false;
+let globalCooldownTimeout: ReturnType<typeof setTimeout> | null = null;
 
 type IconName = keyof typeof LucideIcons;
 
@@ -31,6 +37,8 @@ export default function SingleButton({
   const CustomIcon = customIcon;
   const TagDelay = 100;
   const { toast } = useToast();
+
+  const { mode } = usePerformanceMode();
 
   // Ensure either icon or customIcon is provided
   if (!Icon && !CustomIcon) {
@@ -68,27 +76,24 @@ export default function SingleButton({
     }
   }, [copiedEmail]);
 
-  const handleClick = () => {
-    // minimal cross-browser copy helper
-    const copyText = async (text: string): Promise<boolean> => {
-      try {
-        if (navigator.clipboard?.writeText && window.isSecureContext) {
-          await navigator.clipboard.writeText(text);
-          return true;
-        }
-        // fallback: execCommand
-        const ta = document.createElement("textarea");
-        ta.value = text;
-        document.body.appendChild(ta);
-        ta.select();
-        const ok = document.execCommand("copy");
-        document.body.removeChild(ta);
-        return ok;
-      } catch {
-        return false;
-      }
-    };
+  // derive debounce duration from performance mode
+  const debounceMs = (() => {
+    switch (mode) {
+      case "high":
+        return 0;
+      case "medium":
+        return 75;
+      case "low":
+        return 150;
+      default:
+        return 0;
+    }
+  })();
+  // leading-edge debounce is implemented using module-level globals shared across instances
 
+  // minimal cross-browser copy helper
+
+  const performClick = () => {
     if (emailAddress) {
       const mailto = `mailto:${emailAddress}`;
 
@@ -117,6 +122,32 @@ export default function SingleButton({
     }
 
     onClick?.();
+  };
+
+  // Leading-edge handler: run immediately, then block further runs for debounceMs
+  const handleClick = () => {
+    if (debounceMs === 0) {
+      performClick();
+      return;
+    }
+
+    if (globalBlocked) {
+      // we're in the cooldown window; ignore this click
+      return;
+    }
+
+    // Enter cooldown and perform the click immediately
+    globalBlocked = true;
+    performClick();
+
+    if (globalCooldownTimeout) {
+      clearTimeout(globalCooldownTimeout);
+    }
+
+    globalCooldownTimeout = setTimeout(() => {
+      globalBlocked = false;
+      globalCooldownTimeout = null;
+    }, debounceMs);
   };
 
   const displayLabel = copiedEmail ? "Email copied!" : label;
